% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bergeys}
\alias{bergeys}
\title{Dataset of genera and abstracts from Bergey's Manual}
\format{
A \code{tibble} with 2 columns, which are
\itemize{
  \item \code{genus}: Bacterial genera
  \item \code{abstract}: Corresponding abstract
}
}
\source{
Web scraped from \url{https://onlinelibrary.wiley.com/browse/book/10.1002/9781118960608/toc}
}
\usage{
bergeys
}
\description{
A \code{tibble} with bacterial genera in the first column
and its corresponding abstracts in the second column.
}
\examples{
data("bergeys", package = "bergeys.webscraper")
head(bergeys)

}
\keyword{datasets}
