% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{id_bergeys}
\alias{id_bergeys}
\title{The bergeys dataset mapped to NCBI taxid}
\format{
A \code{data.frame} with 3 columns, which are
\itemize{
   \item \code{ncbi_id}: NCBI taxonomy ID
   \item \code{genus}: Bacterial genera
   \item \code{abstract}: Abstract
}
}
\usage{
id_bergeys
}
\description{
A \code{data.frame} with NCBI taxonomy IDs in the
first column, genera in the second column, and bergey's abstract
in the third column.
}
\examples{
data("id_bergeys", package = "bergeys.webscraper")
head(id_bergeys)
}
\keyword{datasets}
