% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBmanual.R
\name{readBmanual}
\alias{readBmanual}
\title{Scrape Bergey's manual}
\usage{
readBmanual(a, b)
}
\arguments{
\item{a, b}{A range of articles to scrape.}
}
\value{
A data.frame with genera in the first column
and abstracts in the second.
}
\description{
This function scrapes the genus and abstract
from articles in Bergey's Manual.
}
